FAR patch 00678.Opt Дата              : 23.05.2001 Сделал            : Oleg Taranenko Описание          : добавление опций Измененные файлы  : config.cpp interf.cpp manager.cpp struct.hpp fileedit.cpp Новые файлы       : Удаленные файлы   : Состав            : 00678.Opt.txt                     config.cpp.678.diff                     interf.cpp.678.diff                     manager.cpp.678.diff                     struct.hpp.678.diff                     fileedit.cpp.678.diff Основан на патче  : 675 Дополнение        :
+ Опция AllowReedit позволяет выбирать между возможность повторного
  открывания файла на редактирование файла
  ---
  [HKEY_CURRENT_USER\Software\Far\Confirmations]
   "AllowReedit":REG_DWORD
   Параметр "Confirmations/AllowReedit" с типом DWORD изменяет поведение
   открытия файла на редактирование если данный файл уже редактируется.
      0 - Если уже открытый файл не был изменен, то происходит переход к
          открытому редактору без дополнительных вопросов. Если файл был
          изменен, то задается вопрос, и в случае если выбрана вариант
          Reload, то загружается новая копия файла, при этом сделанные
          изменения теряются.
      1 - Так как было раньше. Задается вопрос и происходит переход либо
          уже к открытому файлу либо загружается новая версия редактора.
   По умолчанию = 1
  ---

+ Opt.AltF9 - позволяет выбирать между старым поведение м AltF9 и продвинутым...
  на весь экран :)
  ---
  [HKEY_CURRENT_USER\Software\Far\Interface]
   "AltF9":REG_DWORD
   Параметр "Interface/AltF9" c типом DWORD позволяет выбрать механизм
   работы комбинации Alt-F9 (Изменение размера экрана) в оконном режиме:
     1 - использовать усовершенствованный механизм - окно FAR Manager
         будет переключаться с нормального на максимально доступный размер
         консольного окна и обратно. Размер шрифта консольного окна
         меняться не будет.
         Замечание для Win9x: если для окна FAR Manager выставлен режим
                              автоматического подбора шрифта, то при
                              переключении в максимальный размер шрифт
                              так же будет меняться.
     0 - использовать механизм, совместимый с FAR версии 1.70 beta 3 и
         ниже, т.е. переключение 25/50 линий.
   Данный параметр влияет только на оконный режим работы FAR Manager.
   По умолчанию действует правило 1.
  ---

- Исправление бага - удаление Frame, не внесенного в список FrameList.
  Очень пригодится впоследствии :)

+++ far/fileedit.cpp
@@ -5,10 +5,14 @@ fileedit.cpp
 
 */
 
-/* Revision: 1.46 19.05.2001 $ */
+/* Revision: 1.47 23.05.2001 $ */
 
 /*
 Modify:
+  23.05.2001 OT
+    + Опция AllowReedit
+    ! Исправление бага из-за которго файл трапался по отмене ReloadAgain
+    - Выпрямление логики вызовов в NFZ
   19.05.2001 DJ
     ! лечим последствия NFZ
   15.05.2001 OT
@@ -196,7 +200,7 @@ void FileEditor::Init(char *Name,int CreateNewFile,int EnableSwitch,
     int FramePos=FrameManager->FindFrameByFile(MODALTYPE_EDITOR,FullFileName);
     if (FramePos!=-1) {
       int SwitchTo=FALSE;
-      if (!(*FrameManager)[FramePos]->GetCanLoseFocus(TRUE)) {
+      if (!(*FrameManager)[FramePos]->GetCanLoseFocus(TRUE)||Opt.Confirm.AllowReedit) {
         int MsgCode=Message(0,2,MSG(MEditTitle),FullFileName,MSG(MAskReload),
           MSG(MCurrent),MSG(MReload));
         switch(MsgCode){
@@ -204,11 +208,16 @@ void FileEditor::Init(char *Name,int CreateNewFile,int EnableSwitch,
           SwitchTo=TRUE;
           break;
         case 1:
-          FrameManager->DeleteFrame(FramePos);
-          SetExitCode(-2);
+          if (Opt.Confirm.AllowReedit){
+            SwitchTo=FALSE;
+//            SetExitCode(-1);
+          }else {
+            FrameManager->DeleteFrame(FramePos);
+            SetExitCode(-2);
+          }
           break;
         default:
-          FrameManager->DeleteFrame();
+          FrameManager->DeleteFrame(this);
           SetExitCode(XC_QUIT);
           return;
         }
